requirements.txt
google-ads
python-dotenv
pandas